+++ Vehlution(Everything)/Vehlution(Everything)/Controllers/USER_ROLEController.cs
@@ -51,13 +51,21 @@ namespace Vehlution_Everything_.Controllers
       
         public ActionResult Create([Bind(Include = "USERROLE_ID,ROLE,BLOCKED")] USER_ROLE uSER_ROLE)
         {
-            if (ModelState.IsValid)
+            try
             {
-                db.USER_ROLE.Add(uSER_ROLE);
-                db.SaveChanges();
+                if (ModelState.IsValid)
+                {
+                    db.USER_ROLE.Add(uSER_ROLE);
+                    db.SaveChanges();
+                    TempData["AlertMessage"] = "A user role has sucessfully been addedd!";
+                    return RedirectToAction("Index");
+                }
+            }
+            catch
+            {
+                TempData["AlertMessage"] = "Sorry something went wrong, please try again later";
                 return RedirectToAction("Index");
             }
-
             return View(uSER_ROLE);
         }
 
@@ -83,12 +91,22 @@ namespace Vehlution_Everything_.Controllers
      
         public ActionResult Edit([Bind(Include = "USERROLE_ID,ROLE,BLOCKED")] USER_ROLE uSER_ROLE)
         {
-            if (ModelState.IsValid)
+            try
             {
-                db.Entry(uSER_ROLE).State = EntityState.Modified;
-                db.SaveChanges();
+                if (ModelState.IsValid)
+                {
+                    db.Entry(uSER_ROLE).State = EntityState.Modified;
+                    db.SaveChanges();
+                    TempData["AlertMessage"] = "A user role has sucessfully been updated!";
+                    return RedirectToAction("Index");
+                }
+            }
+            catch
+            {
+                TempData["AlertMessage"] = "Sorry something went wrong, please try again later";
                 return RedirectToAction("Index");
             }
+
             return View(uSER_ROLE);
         }
 
@@ -109,14 +127,33 @@ namespace Vehlution_Everything_.Controllers
 
         // POST: USER_ROLE/Delete/5
         [HttpPost, ActionName("Delete")]
-    
+
         public ActionResult DeleteConfirmed(int id)
         {
-            USER_ROLE uSER_ROLE = db.USER_ROLE.Find(id);
-            db.USER_ROLE.Remove(uSER_ROLE);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            try
+            {
+                USER uSER = new USER();
+                USER_ROLE uSER_ROLE = db.USER_ROLE.Find(id);
+                uSER = db.USERs.Where(zz => zz.USERROLE_ID == uSER_ROLE.USERROLE_ID).FirstOrDefault();
+                if (uSER != null)
+                {
+                    TempData["AlertMessage"] = "You can not delete this user role because it is linked to a user in the User table of the database.";
+                    return RedirectToAction("Index");
+                }
+                db.USER_ROLE.Remove(uSER_ROLE);
+                db.SaveChanges();
+                TempData["AlertMessage"] = "A user role has sucessfully been deleted!";
+                return RedirectToAction("Index");
+            }
+
+            catch
+            {
+                TempData["AlertMessage"] = "Sorry something went wrong, please try again later";
+                return RedirectToAction("Index");
+            }
         }
+
+
 
         protected override void Dispose(bool disposing)
         {
